Remove custom logic
