+++ src/app/grid/grid-paging-template-sample/remote-paging-sample.component.html
@@ -1,0 +1,42 @@
+<div class="grid__wrapper">
+    <igx-grid #grid1 [data]="data | async" width="100%" height="540px" [paging]="true" [perPage]="perPage" [paginationTemplate]="customPager">
+        <igx-column field="ID"></igx-column>
+        <igx-column field="ProductName"></igx-column>
+        <igx-column field="QuantityPerUnit"></igx-column>
+        <igx-column field="SupplierName"></igx-column>
+        <igx-column field="UnitsInStock"></igx-column>
+        <igx-column field="Rating"></igx-column>
+    </igx-grid>
+    <ng-template #customPager let-api>
+        <igx-paginator #paginator
+            [totalRecords]="totalCount"
+            [(perPage)]="perPage"
+            [selectLabel]="'Records per page:'"
+            [selectOptions]="selectOptions"
+            [displayDensity]="grid1.displayDensity"
+            (pageChange)="paginate($event)">
+        </igx-paginator>
+    </ng-template>
+    <ng-template #secCustomPager let-api>
+        <div class="igx-grid__footer">
+            <div id="numberPager" class="igx-paginator" style="justify-content: center;">
+                <button [disabled]="firstPage" (click)="previousPage()" igxButton="flat">
+                    PREV
+                </button>
+                <span *ngIf="shouldShowFirstPage" (click)="paginate(0)">
+                    <a class="pageNavLinks" [routerLink]=''>{{1}}</a> <span class="pageNavLinks">...</span>
+                </span>
+                <span *ngFor="let item of pages" (click)="paginate(item)">
+                    <a class="pageNavLinks {{activePage(item)}}" [routerLink]=''>{{item + 1}}</a>
+                </span>
+                <span *ngIf="shouldShowLastPage" (click)="paginate(totalPages - 1)">
+                    <span class="pageNavLinks">...</span> <a class="pageNavLinks" [routerLink]=''>{{ totalPages }}</a>
+                </span>
+                <button [disabled]="lastPage" (click)="nextPage()" igxButton="flat">
+                    NEXT
+                </button>
+            </div>
+        </div>
+    </ng-template>
+    <button (click)="changeTemplate()" igxButton="flat" style="margin-top: 10px">Change Paging Template</button>
+</div>
