+++ src/app/grid/grid-remote-paging-sample/remote-paging-sample.component.html
@@ -17,26 +17,4 @@
             (pageChange)="paginate($event)">
         </igx-paginator>
     </ng-template>
-    <ng-template #secCustomPager let-api>
-        <div class="igx-grid__footer">
-            <div id="numberPager" class="igx-paginator" style="justify-content: center;">
-                <button [disabled]="firstPage" (click)="previousPage()" igxButton="flat">
-                    PREV
-                </button>
-                <span *ngIf="shouldShowFirstPage" (click)="paginate(0)">
-                    <a class="pageNavLinks" [routerLink]=''>{{1}}</a> <span class="pageNavLinks">...</span>
-                </span>
-                <span *ngFor="let item of pages" (click)="paginate(item)">
-                    <a class="pageNavLinks {{activePage(item)}}" [routerLink]=''>{{item + 1}}</a>
-                </span>
-                <span *ngIf="shouldShowLastPage" (click)="paginate(totalPages - 1)">
-                    <span class="pageNavLinks">...</span> <a class="pageNavLinks" [routerLink]=''>{{ totalPages }}</a>
-                </span>
-                <button [disabled]="lastPage" (click)="nextPage()" igxButton="flat">
-                    NEXT
-                </button>
-            </div>
-        </div>
-    </ng-template>
-    <button (click)="changeTemplate()" igxButton="flat" style="margin-top: 10px">Change Paging Template</button>
 </div>
